Add ratings to TrackDetailed

+++ pkg/types/track.go
@@ -33,7 +33,7 @@ type TrackDetailed struct {
 	MediaFile     *MediaFile `db:"media_file" json:"media_file"`
 	PlayCount     int        `db:"play_count" json:"play_count"`
 	ContextID     *uuid.UUID `db:"context_id" json:"context_id" ts_type:"string"`
-	Rating        *int       `json:"rating"`
+	Rating        *float64   `json:"rating"`
 	UserRating    *int       `json:"user_rating"`
 	CreatedAt     time.Time  `db:"created_at" json:"created_at"`
 	UpdatedAt     time.Time  `db:"updated_at" json:"updated_at"`
